Refactor: Extract the Windows-specific code to a separate file
